# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Lead Calling & Qualification System API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Register a new user
 */
export const registerUserBodyPasswordMin = 6;

export const RegisterUserBody = zod.object({
  name: zod.string(),
  email: zod.string().email(),
  password: zod.string().min(registerUserBodyPasswordMin),
});

/**
 * @summary Login
 */
export const LoginUserBody = zod.object({
  email: zod.string().email(),
  password: zod.string(),
});

export const LoginUserResponse = zod.object({
  token: zod.string(),
  user: zod.object({
    id: zod.number(),
    name: zod.string(),
    email: zod.string(),
  }),
});

/**
 * @summary Create a lead
 */
export const CreateLeadBody = zod.object({
  name: zod.string(),
  phone: zod.string(),
  source: zod.string().optional(),
  notes: zod.string().optional(),
});

/**
 * @summary List leads
 */
export const getLeadsQueryLimitDefault = 50;
export const getLeadsQueryOffsetDefault = 0;

export const GetLeadsQueryParams = zod.object({
  status: zod
    .enum([
      "pending",
      "calling",
      "completed",
      "interested",
      "not_interested",
      "no_response",
    ])
    .optional(),
  search: zod.coerce.string().optional(),
  limit: zod.coerce.number().default(getLeadsQueryLimitDefault),
  offset: zod.coerce.number().default(getLeadsQueryOffsetDefault),
});

export const GetLeadsResponse = zod.object({
  leads: zod.array(
    zod.object({
      id: zod.number(),
      name: zod.string(),
      phone: zod.string(),
      source: zod.string().nullish(),
      status: zod.enum([
        "pending",
        "calling",
        "completed",
        "interested",
        "not_interested",
        "no_response",
      ]),
      retryCount: zod.string(),
      notes: zod.string().nullish(),
      createdAt: zod.coerce.date(),
      updatedAt: zod.coerce.date(),
    }),
  ),
  count: zod.number(),
});

/**
 * @summary Get lead by ID
 */
export const GetLeadByIdParams = zod.object({
  id: zod.coerce.number(),
});

export const GetLeadByIdResponse = zod.object({
  lead: zod.object({
    id: zod.number(),
    name: zod.string(),
    phone: zod.string(),
    source: zod.string().nullish(),
    status: zod.enum([
      "pending",
      "calling",
      "completed",
      "interested",
      "not_interested",
      "no_response",
    ]),
    retryCount: zod.string(),
    notes: zod.string().nullish(),
    createdAt: zod.coerce.date(),
    updatedAt: zod.coerce.date(),
  }),
});

/**
 * @summary Manually initiate a call for a lead
 */
export const InitiateCallParams = zod.object({
  leadId: zod.coerce.number(),
});

export const InitiateCallResponse = zod.object({
  message: zod.string(),
  leadId: zod.number(),
});

/**
 * @summary List all calls
 */
export const getCallsQueryLimitDefault = 50;
export const getCallsQueryOffsetDefault = 0;

export const GetCallsQueryParams = zod.object({
  status: zod.coerce.string().optional(),
  limit: zod.coerce.number().default(getCallsQueryLimitDefault),
  offset: zod.coerce.number().default(getCallsQueryOffsetDefault),
});

export const GetCallsResponse = zod.object({
  calls: zod.array(
    zod.object({
      id: zod.number(),
      leadId: zod.number(),
      twilioCallSid: zod.string().nullish(),
      callStatus: zod.enum([
        "initiated",
        "ringing",
        "answered",
        "completed",
        "no-answer",
        "busy",
        "failed",
      ]),
      duration: zod.number().nullish(),
      recordingUrl: zod.string().nullish(),
      transcript: zod.string().nullish(),
      createdAt: zod.coerce.date(),
      updatedAt: zod.coerce.date(),
    }),
  ),
  count: zod.number(),
});

/**
 * @summary Get call by ID
 */
export const GetCallByIdParams = zod.object({
  id: zod.coerce.number(),
});

export const GetCallByIdResponse = zod.object({
  call: zod.object({
    id: zod.number(),
    leadId: zod.number(),
    twilioCallSid: zod.string().nullish(),
    callStatus: zod.enum([
      "initiated",
      "ringing",
      "answered",
      "completed",
      "no-answer",
      "busy",
      "failed",
    ]),
    duration: zod.number().nullish(),
    recordingUrl: zod.string().nullish(),
    transcript: zod.string().nullish(),
    createdAt: zod.coerce.date(),
    updatedAt: zod.coerce.date(),
  }),
});

/**
 * @summary Get dashboard statistics
 */
export const GetDashboardStatsResponse = zod.object({
  leads: zod.object({
    total: zod.number(),
    byStatus: zod.record(zod.string(), zod.number()),
    recent: zod.array(
      zod.object({
        id: zod.number(),
        name: zod.string(),
        phone: zod.string(),
        source: zod.string().nullish(),
        status: zod.enum([
          "pending",
          "calling",
          "completed",
          "interested",
          "not_interested",
          "no_response",
        ]),
        retryCount: zod.string(),
        notes: zod.string().nullish(),
        createdAt: zod.coerce.date(),
        updatedAt: zod.coerce.date(),
      }),
    ),
  }),
  calls: zod.object({
    total: zod.number(),
    byStatus: zod.record(zod.string(), zod.number()),
    recent: zod.array(
      zod.object({
        id: zod.number(),
        leadId: zod.number(),
        twilioCallSid: zod.string().nullish(),
        callStatus: zod.enum([
          "initiated",
          "ringing",
          "answered",
          "completed",
          "no-answer",
          "busy",
          "failed",
        ]),
        duration: zod.number().nullish(),
        recordingUrl: zod.string().nullish(),
        transcript: zod.string().nullish(),
        createdAt: zod.coerce.date(),
        updatedAt: zod.coerce.date(),
      }),
    ),
  }),
  queue: zod.object({
    total: zod.number(),
    pending: zod.number(),
    scheduled: zod.number(),
  }),
});

/**
 * @summary Trigger AI analysis for a call
 */
export const AnalyzeCallParams = zod.object({
  callId: zod.coerce.number(),
});

export const AnalyzeCallResponse = zod.object({
  message: zod.string(),
  callId: zod.number(),
});